fix(flake-parts): Do not require terranix to be in root of inputs
This change removes the reference to `inputs.terranix` used in the flake
module, so the module can be used also in flake-parts based flakes that
do not import terranix directly, but only as a transitive dependency.

Now it is irelevant, where the terranix flake resides in inputs, whether
it is at `inputs.terranix` (as required now) or e.g. as
`inputs.some-flake.inputs.terranix.